requirements.txt
flask
rembg
Pillow
gunicorn
